Rename `VirtualMonoRepo.E2E.Tests` to `Codeflow.Tests`
