+++ arcade-services.sln
@@ -40,8 +40,6 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "test", "test", "{1A456CF0-C
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Microsoft.DotNet.DarcLib.Tests", "test\Microsoft.DotNet.DarcLib.Tests\Microsoft.DotNet.DarcLib.Tests.csproj", "{6F7DE404-687A-4E0F-B1C3-87AA9D614148}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Microsoft.DotNet.Darc.VirtualMonoRepo.E2E.Tests", "test\Microsoft.DotNet.Darc.VirtualMonoRepo.E2E.Tests\Microsoft.DotNet.Darc.VirtualMonoRepo.E2E.Tests.csproj", "{02959294-F47F-437B-8482-A0E6DC3FBD27}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Microsoft.DotNet.Darc.Tests", "test\Microsoft.DotNet.Darc.Tests\Microsoft.DotNet.Darc.Tests.csproj", "{A1C84AC6-8E21-4B6B-AB82-D17D01D3B0DC}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Maestro.Data.Tests", "test\Maestro.Data.Tests\Maestro.Data.Tests.csproj", "{0BB9EE0F-7EDA-4875-AE72-9A2C3C65DEA3}"
@@ -95,6 +93,8 @@ EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "VersionPropsFormatter", "tools\VersionPropsFormatter\VersionPropsFormatter.csproj", "{40B3EEE3-2B38-4264-A32F-173D129090F8}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "VersionPropsFormatter.Tests", "test\VersionPropsFormatter.Tests\VersionPropsFormatter.Tests.csproj", "{9B43D741-AED1-42DA-88EA-19CB1D71C41C}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Microsoft.DotNet.DarcLib.Codeflow.Tests", "test\Microsoft.DotNet.DarcLib.Codeflow.Tests\Microsoft.DotNet.DarcLib.Codeflow.Tests.csproj", "{A0F47A97-6763-8087-7359-CE0C4E5CF09D}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -202,18 +202,6 @@ Global
 		{6F7DE404-687A-4E0F-B1C3-87AA9D614148}.Release|x64.Build.0 = Release|Any CPU
 		{6F7DE404-687A-4E0F-B1C3-87AA9D614148}.Release|x86.ActiveCfg = Release|Any CPU
 		{6F7DE404-687A-4E0F-B1C3-87AA9D614148}.Release|x86.Build.0 = Release|Any CPU
-		{02959294-F47F-437B-8482-A0E6DC3FBD27}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{02959294-F47F-437B-8482-A0E6DC3FBD27}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{02959294-F47F-437B-8482-A0E6DC3FBD27}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{02959294-F47F-437B-8482-A0E6DC3FBD27}.Debug|x64.Build.0 = Debug|Any CPU
-		{02959294-F47F-437B-8482-A0E6DC3FBD27}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{02959294-F47F-437B-8482-A0E6DC3FBD27}.Debug|x86.Build.0 = Debug|Any CPU
-		{02959294-F47F-437B-8482-A0E6DC3FBD27}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{02959294-F47F-437B-8482-A0E6DC3FBD27}.Release|Any CPU.Build.0 = Release|Any CPU
-		{02959294-F47F-437B-8482-A0E6DC3FBD27}.Release|x64.ActiveCfg = Release|Any CPU
-		{02959294-F47F-437B-8482-A0E6DC3FBD27}.Release|x64.Build.0 = Release|Any CPU
-		{02959294-F47F-437B-8482-A0E6DC3FBD27}.Release|x86.ActiveCfg = Release|Any CPU
-		{02959294-F47F-437B-8482-A0E6DC3FBD27}.Release|x86.Build.0 = Release|Any CPU
 		{A1C84AC6-8E21-4B6B-AB82-D17D01D3B0DC}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{A1C84AC6-8E21-4B6B-AB82-D17D01D3B0DC}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{A1C84AC6-8E21-4B6B-AB82-D17D01D3B0DC}.Debug|x64.ActiveCfg = Debug|Any CPU
@@ -514,6 +502,18 @@ Global
 		{9B43D741-AED1-42DA-88EA-19CB1D71C41C}.Release|x64.Build.0 = Release|Any CPU
 		{9B43D741-AED1-42DA-88EA-19CB1D71C41C}.Release|x86.ActiveCfg = Release|Any CPU
 		{9B43D741-AED1-42DA-88EA-19CB1D71C41C}.Release|x86.Build.0 = Release|Any CPU
+		{A0F47A97-6763-8087-7359-CE0C4E5CF09D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{A0F47A97-6763-8087-7359-CE0C4E5CF09D}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{A0F47A97-6763-8087-7359-CE0C4E5CF09D}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{A0F47A97-6763-8087-7359-CE0C4E5CF09D}.Debug|x64.Build.0 = Debug|Any CPU
+		{A0F47A97-6763-8087-7359-CE0C4E5CF09D}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{A0F47A97-6763-8087-7359-CE0C4E5CF09D}.Debug|x86.Build.0 = Debug|Any CPU
+		{A0F47A97-6763-8087-7359-CE0C4E5CF09D}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{A0F47A97-6763-8087-7359-CE0C4E5CF09D}.Release|Any CPU.Build.0 = Release|Any CPU
+		{A0F47A97-6763-8087-7359-CE0C4E5CF09D}.Release|x64.ActiveCfg = Release|Any CPU
+		{A0F47A97-6763-8087-7359-CE0C4E5CF09D}.Release|x64.Build.0 = Release|Any CPU
+		{A0F47A97-6763-8087-7359-CE0C4E5CF09D}.Release|x86.ActiveCfg = Release|Any CPU
+		{A0F47A97-6763-8087-7359-CE0C4E5CF09D}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -529,7 +529,6 @@ Global
 		{5F9AF839-DB2A-4EFB-94A0-4DE3D4FCDC33} = {C0B90032-033F-4C3E-809F-BD3339E1E313}
 		{A5A48DE5-6577-4F7B-B1F9-39694A14325E} = {C0B90032-033F-4C3E-809F-BD3339E1E313}
 		{6F7DE404-687A-4E0F-B1C3-87AA9D614148} = {1A456CF0-C09A-4DE6-89CE-1110EED31180}
-		{02959294-F47F-437B-8482-A0E6DC3FBD27} = {1A456CF0-C09A-4DE6-89CE-1110EED31180}
 		{A1C84AC6-8E21-4B6B-AB82-D17D01D3B0DC} = {1A456CF0-C09A-4DE6-89CE-1110EED31180}
 		{0BB9EE0F-7EDA-4875-AE72-9A2C3C65DEA3} = {1A456CF0-C09A-4DE6-89CE-1110EED31180}
 		{C6A9F0BA-8C45-44D6-A328-01DE776177DE} = {1A456CF0-C09A-4DE6-89CE-1110EED31180}
@@ -556,6 +555,7 @@ Global
 		{977060F4-EEEF-420D-02E8-5E58D3E51942} = {02EA681E-C7D8-13C7-8484-4AC65E1B71E8}
 		{40B3EEE3-2B38-4264-A32F-173D129090F8} = {02EA681E-C7D8-13C7-8484-4AC65E1B71E8}
 		{9B43D741-AED1-42DA-88EA-19CB1D71C41C} = {1A456CF0-C09A-4DE6-89CE-1110EED31180}
+		{A0F47A97-6763-8087-7359-CE0C4E5CF09D} = {1A456CF0-C09A-4DE6-89CE-1110EED31180}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {32B9C883-432E-4FC8-A1BF-090EB033DD5B}
