+++ test/Microsoft.DotNet.DarcLib.Codeflow.Tests/Resources/example.patch
