+++ test/Microsoft.DotNet.DarcLib.Codeflow.Tests/Resources/new-patch.patch
