+++ test/Microsoft.DotNet.DarcLib.Codeflow.Tests/Resources/remove-file.patch
