+++ test/Microsoft.DotNet.DarcLib.Codeflow.Tests/Resources/test-file-after-changed-patch.txt
