+++ test/Microsoft.DotNet.DarcLib.Codeflow.Tests/Resources/test-file-after-new-patch.txt
