+++ test/Microsoft.DotNet.DarcLib.Codeflow.Tests/Resources/test-file-after-patch.txt
