+++ test/Microsoft.DotNet.DarcLib.Codeflow.Tests/VmrCodeFlowTests.cs
@@ -14,7 +14,7 @@ using Microsoft.DotNet.ProductConstructionService.Client.Models;
 using Microsoft.Extensions.DependencyInjection;
 using NUnit.Framework;
 
-namespace Microsoft.DotNet.Darc.VirtualMonoRepo.E2E.Tests;
+namespace Microsoft.DotNet.DarcLib.Codeflow.Tests;
 
 internal abstract class VmrCodeFlowTests : VmrTestsBase
 {
