+++ test/Microsoft.DotNet.DarcLib.Codeflow.Tests/VmrForwardFlowTest.cs
@@ -13,7 +13,7 @@ using Microsoft.DotNet.DarcLib.VirtualMonoRepo;
 using Microsoft.Extensions.DependencyInjection;
 using NUnit.Framework;
 
-namespace Microsoft.DotNet.Darc.VirtualMonoRepo.E2E.Tests;
+namespace Microsoft.DotNet.DarcLib.Codeflow.Tests;
 
 [TestFixture]
 internal class VmrForwardFlowTest : VmrCodeFlowTests
