+++ test/Microsoft.DotNet.DarcLib.Codeflow.Tests/VmrPatchAddingFileTest.cs
@@ -5,30 +5,39 @@ using System.Collections.Generic;
 using System.IO;
 using System.Threading.Tasks;
 using Microsoft.DotNet.DarcLib.Helpers;
+using Microsoft.DotNet.DarcLib.VirtualMonoRepo;
 using NUnit.Framework;
 
 
-namespace Microsoft.DotNet.Darc.VirtualMonoRepo.E2E.Tests;
+namespace Microsoft.DotNet.DarcLib.Codeflow.Tests;
 
 [TestFixture]
-internal class VmrPatchRemovingFileTest : VmrPatchesTestsBase
+internal class VmrPatchAddingFileTest : VmrPatchesTestsBase
 {
-    public VmrPatchRemovingFileTest() : base("remove-file.patch")
+    private readonly string _productRepoNewFile = "new-file.txt";
+
+    public VmrPatchAddingFileTest() : base("add-file.patch")
     {
     }
 
     [Test]
     public async Task VmrPatchAddsFileTest()
     {
+        var vmrSourcesPath = VmrPath / VmrInfo.SourcesDir;
         var patchPathInRepo = InstallerPatchesDir / PatchFileName;
 
         await InitializeRepoAtLastCommit(Constants.InstallerRepoName, InstallerRepoPath);
         await InitializeRepoAtLastCommit(Constants.ProductRepoName, ProductRepoPath);
 
+        var newFilePath = vmrSourcesPath / Constants.ProductRepoName / _productRepoNewFile;
+        var patchPath = VmrPatchesDir / PatchFileName;
+
         var expectedFilesFromRepos = new List<NativePath>
         {
-            VmrPatchesDir / PatchFileName,
-            InstallerFilePathInVmr
+            ProductRepoFilePathInVmr,
+            InstallerFilePathInVmr,
+            newFilePath,
+            patchPath
         };
 
         var expectedFiles = GetExpectedFilesInVmr(
@@ -42,10 +51,9 @@ internal class VmrPatchRemovingFileTest : VmrPatchesTestsBase
         await GitOperations.CommitAll(InstallerRepoPath, "Remove the patch file");
         await UpdateRepoToLastCommit(Constants.InstallerRepoName, InstallerRepoPath);
 
-        expectedFiles.Add(ProductRepoFilePathInVmr);
-        expectedFiles.Remove(VmrPatchesDir / PatchFileName);
+        expectedFiles.Remove(newFilePath);
+        expectedFiles.Remove(patchPath);
 
         CheckDirectoryContents(VmrPath, expectedFiles);
     }
 }
-
