+++ test/Microsoft.DotNet.DarcLib.Codeflow.Tests/VmrPatchChangingFileTest.cs
@@ -9,7 +9,7 @@ using FluentAssertions;
 using Microsoft.DotNet.DarcLib.Helpers;
 using NUnit.Framework;
 
-namespace Microsoft.DotNet.Darc.VirtualMonoRepo.E2E.Tests;
+namespace Microsoft.DotNet.DarcLib.Codeflow.Tests;
 
 [TestFixture]
 internal class VmrPatchChangingFileTest : VmrPatchesTestsBase
