+++ test/Microsoft.DotNet.DarcLib.Codeflow.Tests/VmrPatchRemovingFileTest.cs
@@ -5,39 +5,30 @@ using System.Collections.Generic;
 using System.IO;
 using System.Threading.Tasks;
 using Microsoft.DotNet.DarcLib.Helpers;
-using Microsoft.DotNet.DarcLib.VirtualMonoRepo;
 using NUnit.Framework;
 
 
-namespace Microsoft.DotNet.Darc.VirtualMonoRepo.E2E.Tests;
+namespace Microsoft.DotNet.DarcLib.Codeflow.Tests;
 
 [TestFixture]
-internal class VmrPatchAddingFileTest : VmrPatchesTestsBase
+internal class VmrPatchRemovingFileTest : VmrPatchesTestsBase
 {
-    private readonly string _productRepoNewFile = "new-file.txt";
-
-    public VmrPatchAddingFileTest() : base("add-file.patch")
+    public VmrPatchRemovingFileTest() : base("remove-file.patch")
     {
     }
 
     [Test]
     public async Task VmrPatchAddsFileTest()
     {
-        var vmrSourcesPath = VmrPath / VmrInfo.SourcesDir;
         var patchPathInRepo = InstallerPatchesDir / PatchFileName;
 
         await InitializeRepoAtLastCommit(Constants.InstallerRepoName, InstallerRepoPath);
         await InitializeRepoAtLastCommit(Constants.ProductRepoName, ProductRepoPath);
 
-        var newFilePath = vmrSourcesPath / Constants.ProductRepoName / _productRepoNewFile;
-        var patchPath = VmrPatchesDir / PatchFileName;
-
         var expectedFilesFromRepos = new List<NativePath>
         {
-            ProductRepoFilePathInVmr,
-            InstallerFilePathInVmr,
-            newFilePath,
-            patchPath
+            VmrPatchesDir / PatchFileName,
+            InstallerFilePathInVmr
         };
 
         var expectedFiles = GetExpectedFilesInVmr(
@@ -51,9 +42,10 @@ internal class VmrPatchAddingFileTest : VmrPatchesTestsBase
         await GitOperations.CommitAll(InstallerRepoPath, "Remove the patch file");
         await UpdateRepoToLastCommit(Constants.InstallerRepoName, InstallerRepoPath);
 
-        expectedFiles.Remove(newFilePath);
-        expectedFiles.Remove(patchPath);
+        expectedFiles.Add(ProductRepoFilePathInVmr);
+        expectedFiles.Remove(VmrPatchesDir / PatchFileName);
 
         CheckDirectoryContents(VmrPath, expectedFiles);
     }
 }
+
