+++ test/Microsoft.DotNet.DarcLib.Codeflow.Tests/VmrPatchesTestsBase.cs
@@ -8,7 +8,7 @@ using Microsoft.DotNet.DarcLib.Models.VirtualMonoRepo;
 using Microsoft.DotNet.DarcLib.VirtualMonoRepo;
 
 
-namespace Microsoft.DotNet.Darc.VirtualMonoRepo.E2E.Tests;
+namespace Microsoft.DotNet.DarcLib.Codeflow.Tests;
 
 internal class VmrPatchesTestsBase : VmrTestsBase
 {
