+++ test/Microsoft.DotNet.DarcLib.Codeflow.Tests/VmrRecursiveSyncTests.cs
@@ -10,7 +10,7 @@ using Microsoft.DotNet.DarcLib.Models.VirtualMonoRepo;
 using Microsoft.DotNet.DarcLib.VirtualMonoRepo;
 using NUnit.Framework;
 
-namespace Microsoft.DotNet.Darc.VirtualMonoRepo.E2E.Tests;
+namespace Microsoft.DotNet.DarcLib.Codeflow.Tests;
 
 internal class VmrRecursiveSyncTests : VmrTestsBase
 {
