+++ test/Microsoft.DotNet.DarcLib.Codeflow.Tests/VmrRepositoryCloningTest.cs
@@ -9,7 +9,7 @@ using Microsoft.DotNet.DarcLib.VirtualMonoRepo;
 using Microsoft.Extensions.DependencyInjection;
 using NUnit.Framework;
 
-namespace Microsoft.DotNet.Darc.VirtualMonoRepo.E2E.Tests;
+namespace Microsoft.DotNet.DarcLib.Codeflow.Tests;
 internal class VmrRepositoryCloningTest : VmrTestsBase
 {
     [Test]
