+++ test/Microsoft.DotNet.DarcLib.Codeflow.Tests/VmrSyncAdditionalMappingsTest.cs
@@ -11,7 +11,7 @@ using Microsoft.DotNet.DarcLib.VirtualMonoRepo;
 using NUnit.Framework;
 
 
-namespace Microsoft.DotNet.Darc.VirtualMonoRepo.E2E.Tests;
+namespace Microsoft.DotNet.DarcLib.Codeflow.Tests;
 
 [TestFixture]
 internal class VmrSyncAdditionalMappingsTest : VmrTestsBase
