+++ test/Microsoft.DotNet.DarcLib.Codeflow.Tests/VmrSyncRepoChangesTest.cs
@@ -10,7 +10,7 @@ using Microsoft.DotNet.DarcLib.Models.VirtualMonoRepo;
 using Microsoft.DotNet.DarcLib.VirtualMonoRepo;
 using NUnit.Framework;
 
-namespace Microsoft.DotNet.Darc.VirtualMonoRepo.E2E.Tests;
+namespace Microsoft.DotNet.DarcLib.Codeflow.Tests;
 
 [TestFixture]
 internal class VmrSyncRepoChangesTest : VmrTestsBase
