+++ test/Microsoft.DotNet.DarcLib.Codeflow.Tests/VmrTestsBase.cs
@@ -21,7 +21,7 @@ using Microsoft.Extensions.Logging;
 using Moq;
 using NUnit.Framework;
 
-namespace Microsoft.DotNet.Darc.VirtualMonoRepo.E2E.Tests;
+namespace Microsoft.DotNet.DarcLib.Codeflow.Tests;
 
 internal abstract class VmrTestsBase
 {
