+++ test/Microsoft.DotNet.DarcLib.Codeflow.Tests/VmrTestsOneTimeSetUp.cs
@@ -9,7 +9,7 @@ using Microsoft.DotNet.DarcLib.Helpers;
 using Microsoft.DotNet.DarcLib.VirtualMonoRepo;
 using NUnit.Framework;
 
-namespace Microsoft.DotNet.Darc.VirtualMonoRepo.E2E.Tests;
+namespace Microsoft.DotNet.DarcLib.Codeflow.Tests;
 
 [SetUpFixture]
 public class VmrTestsOneTimeSetUp
